+++ prover/crates/bin/witness_generator/src/recursion_tip/mod.rs
@@ -1,0 +1,2 @@
+pub mod recursion_tip;
+pub use recursion_tip::*;
