+++ prover/crates/bin/witness_generator/src/scheduler/mod.rs
@@ -1,0 +1,2 @@
+pub mod scheduler;
+pub use scheduler::*;
